Fix points cleanup counter for duplicate points.
